+++ app.R
@@ -1,3 +1,3 @@
-# This is file is used by the dev/deploy.R script.
+# This file is used by the dev/deploy.R script.
 pkgload::load_all(export_all = FALSE, helpers = FALSE, attach_testthat = FALSE)
 iaoreader::run_shiny_app(run_app = FALSE)